// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const SenderType = {
  USER: 'USER',
  CHANNEL: 'CHANNEL',
  BOT: 'BOT'
} as const

export type SenderType = (typeof SenderType)[keyof typeof SenderType]


export const ChannelType = {
  PUBLIC: 'PUBLIC',
  PRIVATE: 'PRIVATE'
} as const

export type ChannelType = (typeof ChannelType)[keyof typeof ChannelType]


export const ConversationType = {
  DIRECT: 'DIRECT',
  GROUP: 'GROUP',
  CHANNEL: 'CHANNEL'
} as const

export type ConversationType = (typeof ConversationType)[keyof typeof ConversationType]


export const ConversationRole = {
  OWNER: 'OWNER',
  ADMIN: 'ADMIN',
  MEMBER: 'MEMBER'
} as const

export type ConversationRole = (typeof ConversationRole)[keyof typeof ConversationRole]


export const GroupType = {
  PUBLIC: 'PUBLIC',
  PRIVATE: 'PRIVATE'
} as const

export type GroupType = (typeof GroupType)[keyof typeof GroupType]


export const FileStatus = {
  PENDING: 'PENDING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED'
} as const

export type FileStatus = (typeof FileStatus)[keyof typeof FileStatus]